refactor: 슬로건, error message 상수화
