+++ src/components/Layout/Introduction.tsx
@@ -1,10 +1,14 @@
 import logoImage from '@/assets/logo.png';
+import {
+  SLOGAN_MESSAGE1,
+  SLOGAN_MESSAGE2,
+} from '@/constants/messages/boardSignal';
 
 const Introduction = () => (
   <div className='flex flex-col items-center'>
     <img src={logoImage} />
     <p className='text-center text-gray-500'>
-      "매치업! 보드게임을 즐기는 사람들을 찾는 최적의 장소."
+      {`${SLOGAN_MESSAGE1} ${SLOGAN_MESSAGE2}`}
     </p>
   </div>
 );
