+++ src/constants/messages/boardSignal.ts
@@ -1,2 +1,5 @@
 export const SIGNAL_TEMPERATURE_MESSAGE =
   '시그널온도는 다른 사용자로부터 받은 리뷰, 신고, 운영자 제재 등을 종합해서 만든 매너 지표입니다.';
+
+export const SLOGAN_MESSAGE1 = '"매치업! 보드게임을 즐기는';
+export const SLOGAN_MESSAGE2 = '사람들을 찾는 최적의 장소."';
