+++ src/constants/messages/error.ts
@@ -1,5 +1,14 @@
-// Form error message
+// Form Error Message
 export const REQUIRED_ERROR_MESSAGE = '필수 입력 항목이에요.';
 export const MIN_LENGTH_ERROR_MESSAGE = '자 이상 입력해야 해요.';
 export const TRIM_ERROR_MESSAGE = '앞뒤 공백 제외';
 export const MAX_LENGTH_ERROR_MESSAGE = '자까지 입력할 수 있어요.';
+
+// Error Alert Full Screen Message
+export const DEFAULT_ERROR_TITLE = '잠시 후 다시 시도해주세요 :(';
+export const DEFAULT_ERROR_MESSAGE = '요청사항을 처리하는데 실패했습니다.';
+export const DEFAULT_BUTTON_TITLE = '새로 고침하기';
+
+// Error Toast Message
+export const ERROR_TOAST_MESSAGE =
+  '요청을 처리할 수 없었습니다.\r\n잠시 후 다시 시도해주세요.';
