+++ src/pages/Login/index.tsx
@@ -1,14 +1,15 @@
 import logoImage from '@/assets/logo.png';
 import TabBar from '@/components/TabBar';
+import {
+  SLOGAN_MESSAGE1,
+  SLOGAN_MESSAGE2,
+} from '@/constants/messages/boardSignal';
 
 import {
   KakaoLoginButton,
   NaverLoginButton,
   SkipLoginButton,
 } from './components/LoginButtons';
-
-const SLOGAN_MESSAGE1 = '"매치업! 보드게임을 즐기는';
-const SLOGAN_MESSAGE2 = '사람들을 찾는 최적의 장소."';
 
 const LoginPage = () => {
   return (
